pyproject.toml
[project]
name = "langchain-chains1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.24",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
